WIP: Initial dataset transformation extension for virtual vector vars
Generates speed and direction variables from detected component vars.

Currently a proof of concept, the results are incomplete and
surely incorrect in many (all?) cases.

Specifics on how to render the directional variable with arrows/barbs
and how to structure requests for and layering of the speed/magnitude
and directional layers are TBD.